# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank RestAPI
  description: Документация API для Bank rest
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: CardController
  description: Работа с картами
- name: AuthController
  description: Контроллер для регистрации и авторизации
- name: TransactionController
  description: Работа с транзакциями
paths:
  /registration:
    post:
      tags:
      - AuthController
      summary: Регистрация
      operationId: registration
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /card/{id}/block:
    post:
      tags:
      - CardController
      description: Блокирует карту
      operationId: blockCard
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /card/transactions:
    post:
      tags:
      - TransactionController
      description: Перевод денег
      operationId: transactions
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionsDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /auth:
    post:
      tags:
      - AuthController
      summary: Авторизация
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JwtRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /admin/registration:
    post:
      tags:
      - admin-controller
      summary: Создает админа
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationAdminDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /admin/card/create:
    post:
      tags:
      - AdminCardController
      description: Создает пользователю карту
      operationId: createCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /admin/card/{id}:
    get:
      tags:
      - AdminCardController
      description: Получает данные карты с поммощью айди
      operationId: getCardById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - AdminCardController
      description: Удаляет карту
      operationId: deleteCard
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    patch:
      tags:
      - AdminCardController
      description: "Изменяет статус карты, если хотите продлить нужно указать дату\
        \ окончания"
      operationId: changeStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardStatusDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /card/{id}:
    get:
      tags:
      - CardController
      description: Получает данные карты
      operationId: getCardById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /card/:
    get:
      tags:
      - CardController
      description: Возвращает карты авторизированного пользователя
      operationId: getMyCards
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /admin/user/{id}:
    get:
      tags:
      - admin-user-controller
      description: Переносит вас в /admin/card/user/
      operationId: getUsersCard
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /admin/user/:
    get:
      tags:
      - admin-user-controller
      description: Получает всех пользователей и админов
      operationId: getUsers
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /admin/card/{id}/user:
    get:
      tags:
      - AdminCardController
      description: Вернет пользователя с помощью айди карты
      operationId: getUserByCardId
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /admin/card/user/{id}:
    get:
      tags:
      - AdminCardController
      description: Вернеет карты пользователя
      operationId: getCardByUserId
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /admin/card/:
    get:
      tags:
      - AdminCardController
      description: Возвращает все карты
      operationId: getCards
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
components:
  schemas:
    RegistrationDTO:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        confirmPassword:
          type: string
      required:
      - confirmPassword
      - password
      - username
    TransactionsDTO:
      type: object
      properties:
        fromId:
          type: integer
          format: int64
        toId:
          type: integer
          format: int64
        description:
          type: string
        amount:
          type: integer
          format: int64
    JwtRequestDTO:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    RegistrationAdminDTO:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        confirmPassword:
          type: string
    CreateCardDTO:
      type: object
      properties:
        username:
          type: string
        expired:
          type: string
          format: date
        balance:
          type: integer
          format: int64
    CardStatusDTO:
      type: object
      properties:
        status:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
        expiryDate:
          type: string
          format: date